requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.8
defusedxml==0.7.1
Django==4.2.14
django-allauth==0.63.6
django-crispy-forms==2.2
django-environ==0.11.2
django-phonenumber-field==8.0.0
idna==3.7
isodate==0.6.1
lxml==5.2.2
oauthlib==3.2.2
phonenumberslite==8.13.40
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
python3-saml==1.16.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
xmlsec==1.3.14
